requirements.txt
Django==1.8.8
django-import-export==0.4.3
oauth2client==1.5.2
identity-toolkit-python-client>=0.1.6
google-api-python-client==1.4.2
psycopg2==2.6.1
gunicorn==19.4.1
django-constance[database]
scipy==0.16.1
numpy==1.10.4
